Added custom iterator for SpatialMap, and fixed bug in layered spatial map custom iterator.  Added unit tests.
